@tnorlund 2025-08-09
Closing this PR as the good ideas have been cherry-picked and merged via PR #303:

✅ **What was merged (from #303)**:
- ARG_MAX protection with retry logic for pip install commands
- S3 upload retries with exponential backoff
- Build validation checks for layer.zip creation
- CloudWatch logging for all CodeBuild projects
- Debug mode flag for verbose logging
- Improved error messages with context
- Type annotations for better IDE support

❌ **What was intentionally not merged**:
- Pure Python zipfile approach (loses build environment isolation)
- Eliminating CodeBuild (removes proper compilation environment for C extensions)
- Local pip installations (creates architecture mismatches)

The CodeBuild-based architecture has been preserved as it properly handles build environments and package compilation. The stability improvements from this PR have been successfully integrated while maintaining the robust infrastructure approach.

Thank you for the initial work that inspired these improvements!